requirements.txt
Flask==3.0.2
pillow
tensorflow-cpu==2.16.1
keras==3.0.5
